requirements.txt
groq>=0.9.0
playwright>=1.44.0
Pillow>=10.0.0
requests>=2.31.0
streamlit>=1.30.0
